ci: Show what size limit jobs exceeded the limit (#13532)
Noticed that it was a bit annoying, when we exceeded the size limit, to
find out why/what has failed.

Now, it should show in the GH comment which specific check has failed,
and what the current limit is:


![image](https://github.com/user-attachments/assets/13495922-ac91-42bf-a398-18bb7b0e9e72)

While I was at this, I also streamlined the code a bit, extracted it a
bit into utils to de-clutter stuff, and fixed/improved some debug
logging issues (e.g. the log group was not correctly closed before, ...)
